Extract AmazonSageMakerBaseExecutor
